KAN-2 progress
Adjusted DB creation file.

+++ db/mongo_basic_db_setup.js
@@ -3,50 +3,71 @@
 // 3 - run js in mongo shell - mongo < setupDatabase.js
 
 // Switch to the desired database
-use mydatabase;
+use tactease_test1;
+
+//should we reset the DB?
+const do_reset = 1;
 
 // RESET DB - to rebuild from scratch
+if(do_reset){
 db.runCommand({ dropDatabase: 1 });
+}
 
-// Insert users
-db.users.insertMany([
-  { personal_number: '1', fullname: 'user1', class: '' },
-  { personal_number: '2', fullname: 'user2', class: '' },
-]);
+// // Insert users
+// db.users.insertMany([
+//   { personal_number: '1', fullname: 'user1', class: '' },
+//   { personal_number: '2', fullname: 'user2', class: '' },
+// ]);
 
 // Insert classes
 db.classes.insertMany([
-  { id: '1', name: 'object1' },
-  { id: '2', name: 'object2' },
+  { id: '1', name: 'class1' },
+  { id: '2', name: 'class2' },
 ]);
 
-//user type 1
-db.users_s.insertMany([
-  { personal_number: '3', fullname: 'user1', class: '', pakal: '' },
-  { personal_number: '4', fullname: 'user2', class: '', pakal: '' },
+//add commanders
+db.commanders.insertMany([
+  { personal_number: '1', fullname: 'com1', class: ''},
+  { personal_number: '2', fullname: 'com2', class: ''},
 ]);
 
-// //user type 2
-// db.users_c.insertMany([
-//   { personal_number: '3', fullname: 'user1', class: ''},
-//   { personal_number: '4', fullname: 'user2', class: ''},
-// ]);
+//add soldiers
+db.soldiers.insertMany([
+  { personal_number: '3', fullname: 'sol1', class: '', pakal: '', constraints: '' },
+  { personal_number: '4', fullname: 'sol2', class: '', pakal: '', constraints: '' },
+  { personal_number: '5', fullname: 'sol3', class: '', pakal: '', constraints: '' },
+  { personal_number: '6', fullname: 'sol4', class: '', pakal: '', constraints: '' },
+]);
+
+db.missions.insertMany([
+  { mission_id: '1', name: 'miss1', start: '2024-03-01-13:00', end: '2024-03-01-15:00', soldiers: ''},
+  { mission_id: '2', name: 'miss2', start: '2024-05-01-15:00', end: '2024-05-01-16:40', soldiers: ''},
+]);
 
 // Update users with references to classes - many to one
 db.users.update(
-  { personal_number: '1' },
+  { personal_number: '3' },
   {
     $set: {
-      class: db.classes.findOne({ name: 'object1' })._id,
+      class: db.classes.findOne({ name: 'class1' })._id,
     },
   }
 );
 
 db.users.update(
-  { personal_number: '2' },
+  { personal_number: '4' },
   {
     $set: {
-      class: db.classes.findOne({ name: 'object2' })._id,
+      class: db.classes.findOne({ name: 'class2' })._id,
+    },
+  }
+);
+
+db.users.update(
+  { personal_number: '5' },
+  {
+    $set: {
+      class: db.classes.findOne({ name: 'class2' })._id,
     },
   }
 );
